Code golf each transitions (#9476)
* chore: improve code size of transitions

* Add changeset
