+++ .changeset/cuddly-pianos-drop.md
@@ -1,0 +1,5 @@
+---
+'svelte': patch
+---
+
+chore: improve bundle code size
